Fix missing utility classes causing compilation errors - created 7 missing utility classes

+++ libunified/src/main/java/com/mpdc4gsr/libunified/ir/view/ZoomableDraggableView.java
@@ -79,7 +79,7 @@ public class ZoomableDraggableView extends View {
         }
         showBitmapHeight = pxBitmapHeight / yscale;
         showBitmapHeightWidth = pxBitmapHeight * originalBitmapWidth / originalBitmapHeight * xscale;
-        showBitmap = BitmapUtils.scaleWithWH(originalBitmap, showBitmapHeightWidth, showBitmapHeight);
+        showBitmap = BitmapUtils.scaleWithWH(originalBitmap, (int) showBitmapHeightWidth, (int) showBitmapHeight);
     }
 
     @Override
